requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
rembg==2.0.50
python-multipart==0.0.6
requests==2.31.0
Pillow==10.1.0 